Move check_user_is_present_and_has_access_to_given_report() function to utils
